scribe_tail: fix infinite loop on log rotate
scribe_tail was using the results from stat() rather than fstat() to
detect file truncation, resulting in a false positive when filename
is changed to point to a new, empty file.  scribe_tail would then
rewind the open file (not filename) and replay the log repeatedly.

This changes scribe_tail to use the results from fstat() on the open
file to detect truncation.